Move args sanity checks to separate file (as many more needed moving forwards)
